backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
pydantic-settings>=2,<3
sqlalchemy>=2.0,<3
aiosqlite>=0.20,<1
